+++ framework/src/sbt-plugin/src/main/scala/PlaySettings.scala
@@ -2,9 +2,7 @@ package sbt
 
 import Keys._
 import PlayKeys._
-
-import java.io.{ FileWriter, Writer }
-import java.util.Properties
+import PlayEclipse._
 
 trait PlaySettings {
   this: PlayCommands with PlayPositionMapper =>
@@ -67,33 +65,6 @@ trait PlaySettings {
     resourceGenerators in Compile <+= LessCompiler,
     resourceGenerators in Compile <+= CoffeescriptCompiler
   )
-
-  def saveScalaIdePlay2Prefs(ref: ProjectRef, structure: Load.BuildStructure, baseDir: File) = {
-    def fileWriterMkdirs(file: File): FileWriter = {
-      file.getParentFile.mkdirs()
-      new FileWriter(file)
-    }
-
-    def saveProperties(file: File, settings: Seq[(String, String)]): Unit =
-      if (!settings.isEmpty) {
-        val properties = new Properties
-        for ((key, value) <- settings) properties.setProperty(key, value)
-        val writer = fileWriterMkdirs(file);
-        try {
-          properties.store(writer, "Generated by sbt-plugin")
-        } finally {
-          writer.close();
-        }
-      }
-
-    Project.getProject(ref, structure).foreach { p =>
-      (templatesImport in ref get structure.data).foreach { imports =>
-        val value = imports.mkString ("import ", "\nimport ", "\n")
-        val properties = Seq(("eclipse.preferences.version", "1"), ("templateImport", value))
-        saveProperties(baseDir / ".settings" / "org.scala-ide.play2.prefs", properties)
-      }
-    }
-  }
 
   lazy val defaultSettings = Seq[Setting[_]](
